Add exclusion regular expressions.
These can be specified in the output string to ignore files (either read
or written) that match the regular expression pattern. Eg:

: |> cmd... |> output.o ^ignore.*bar.txt

For any file that matches the regular expression "ignore.*bar.txt", tup
will ignore it as an input, and will not raise an error for it as an
output. Note that the leading "^" is to signify an exclusion, and is not
part of the regular expression (ie: it doesn't match the start of the
filename string).

Internally, the regular expressions are all stored as strings in the
virtual '^' directory. They are shared amongst commands, so if every
command has the same exclusion string, the database impact should be
negligible.

In order to avoid issues with future scans, tup will dynamically create
nodes in the db for ignored files that are created by a subprocess.
Otherwise tup would see them on a future scan and produce potentially
spurious "Parsing Tupfiles..." messages.