Remove some hard-coded colors.
I usually redirect the sql debug to a file anyway, so colors there
aren't very helpful. And in other cases they are printed to a temporary
file before being displayed under the job banner, and so we don't know
whether or not colors will be enabled for that output yet when we are
adding them to the string.