Added t4205 - make sure we get all the mtimes for full_deps.
If we added a file like '/usr/include/foo.h' when doing full_deps, we
may create '/usr' and /usr/include' as ghost directories with an
mtime=-1. If we later process '/usr/include' as a separate access event,
the tent already exists, so we don't set the mtime. Then on the next
update, we see that the mtime differs and re-trigger compilation on
things unnecessarily.

The fix is to set mtimes when we create ghost directories with full_deps
set.