Added t4208, t4209 - allow external inputs when full_deps=1
If we're running with full_deps enabled, we should be able to list an
external input file and have things work properly. We can do this by
just adding the external path_list to the name_list without a
corresponding tent. An external path_list is differentiated because it
has pl->dt == -1.