Ignore files in /var/run and /run
When running node scripts, node seems to call getaddrinfo(), which
results in resolv.conf being read. On systemd systems, this can point to
a file in /run, which may be updated randomly by the system and cause
spurious rebuilds. We should probably just ignore files in /run and
/var/run, since they should just be used by servers and aren't
persistent across reboots.

We also need to special case /etc/resolv.conf, since that can symlink to
../run/resolv...