Added t4212 - test for running a program compiled with -pg
When compiled with -pg, the program will write to gmon.out. This happens
in an atexit() routine called _mcleanup in glibc, but glibc uses a
direct syscall to write the file, so the ldpreload hook misses the open.
With a FUSE server we would see this write, so it is difficult to write
an accurate Tupfile if the fuse server would need gmon.out in the
outputs and the ldpreload server would not, so we need some way to track
this. It seems that hooking _mcleanup() does the job.

Another option might be to ignore reads/writes to gmon.out, but that
would make it difficult to add a step to run gprof in the Tupfile, which
may be desirable (ie: run a program to generate a new gmon.out, and then
have tup run gprof to produce a readable output from it).