Fixes building and testing in Freebsd
The tupfile changes are trivial. I had to link clang to gcc to get the
tests to work. Three tests (4068, 6052, and 6053) that should work
still do not for reasons I don't understand. Test 4166 arguably should
be made to work, but it's disabled for Mac OS X so I'm leaving it for
now. A few others needed minor fixes.