Added t4213 - case sensitvity Windows tests.
When renaming or unlinking a file with a different case style than was
used to create the file, tup wasn't correctly accounting for that due to
using strcmp() instead of name_cmp(). This would cause clang to fail if
run in MinGW, since some file accesses there would have "c:" and others
"C:". (It appears that GetCurrentDirectory() differs in MinGW and
Cygwin, in that the former has a lower-case c: and the latter an
upper-case C:, so clang could work in one but not the other).

Related but somewhat separate, 'mv' in Cygwin potentially uses
FileRenameInformationEx in NtSetInformationFile instead of always using
FileRenameInformation. This was found by perusing the source for
newlib-cygwin/winsup/cygwin/syscalls.cc, though an updated ProcMon
would've also shown the correct operation. FileRenameInformationEx isn't
yet defined in the MinGW headers, so we need to add it manually.