Added tent_tree structure, currently used for outputs.
In most cases we want the tup_entry, rather than just a tupid, which
means we have to call tup_entry_add() to search through the main
tup_entry root every time we want to do something with a tupid. We can
use a tent_tree structure instead, which is the same size as a
tupid_tree, but gives us the tent directly so we don't need to search
for it.