Use a tup_entry instead of tupid_t dt for node_select / insert
In most places we already have the tup_entry, so going back to a tupid_t
that we have to use to search through the entry tree to find the
original tup_entry is silly. In some cases this makes it a little more
annoying, but those are likely just cases that should also be converted
to tup_entry instead of a tupid.