Work-around issue with !tup_preserve and same inputs/outputs.
Since !tup_preserve copies a file of the same name from the source
directory to the variant build directory, searching for the input
unfortunately grabs the variant copy on a future update. This causes tup
to think that the rule is both reading from and writing the same file
(eg: build/file.txt) rather than reading from file.txt and writing to
build/file.txt.

A simple work-around is to about adding inputs to !tup_preserve, since
all it is doing is a file copy on a source file that we don't need a
sticky link for anyway.

Fixes #388