Update t8092 and add t8093, t8094, t8095 - more !tup_preserve fixes.
The previous fix for #388, while it did report success in parsing the
Tupfile, unfortunately parsed it incorrectly. Specifically, it was still
finding file.txt from the variant dir, and using that to calculate the
expansion of the !tup_preserve command, meaning that the file it pointed
to differed in the first and second run. Simply touching the Tupfile
would cause the command string to change and the preserve commands to
re-execute with the new (wrong) file. t8092 has been updated to check
for this.

Additionally, t8093 was created to make sure that we can use the
preserved file in a subsequent command, which will be pulling from the
variant directory. t8094 does the same thing as t8093 with Lua, and
t8095 tests that !tup_preserve fails if trying to preserve a generated
file.