win32: Use MinGW's pthreads rather than our own custom wrapper.
This was not available in earlier MinGW versions, so we had to roll our
own. Now that it is, we can get rid of win32/pthread.c. Apparently with
the dir_mutex we were relying on it being re-entrant since it was also
used to lock the dirpath. However, that should use its own lock since it
is just locking the dirpath data structure, rather than the cwd of the
thread.