Add ref-counting for tup_entrys in tent_list / tent_tree.
This should be a helpful sanity check going forward. It did catch one
issue where a generated node becomes a normal node when variants are
used, though it didn't seem to be a runtime problem because the sticky
link gets removed from the db when the generated node is deleted.