Use OSX's build-in *at functions.
Now that OSX supports the *at functions for accessing files with
pthreads natively, we don't need to use our wrapper functions.

The wrappers worked until OSX 10.15, when tup would hang trying to
initialize the FUSE server. A call to a *at() function uses the
dir_mutex wrapper to lock the cwd and then fchdir() to the fd for the
*at function. For reasons that still aren't clear to me, the fchdir()
itself blocks for several seconds until it times out (and oddly
returns success). The problem can be reproduced in the fusexmp_fh
example filesystem, with the following conditions:

 1) Add an fchdir() to a function like access()
 2) Run fuse_main() in a thread
 3) Try to access a file in the fuse fs from the main thread

So it seems 10.15 added some process-level locking to filesystems,
possibly specifically to fchdir(), that prevents the *at wrappers from
working. Fortunately, using the built-in functions that are now available
fixes the issue.

Fixes #395