win32: Use NUL for stdin.
Similar to other platforms, we need to make sure stdin is a valid
handle, since some programs may rely on stdin. In particular, the
version of bash in WSL tries to use stdin even without doing a read
(just a 'false | true' would use stdin), and it would complain that "The
paramter is incorrect". (Oddly, the message is in wchars, so the message
prints out as just "T", but the full message can be seen in the output
file). Using INVALID_HANDLE_VALUE also doesn't work, but using a NUL
file handle does.

Note that the WSL bash can get picked up by running tests in tup if tup
is compiled as 64-bit, because C:/Windows/System32/bash.exe is in the
PATH which does some launcher lookup stuff to find the WSL bash.

This also fixes a minor issue of missing end-quotes with bash, but bash
seems to process the command-line correctly anyway.