Fix 'tup version' in bootstrapped tup.
tup_version was changed from a function to a global variable a while
ago, but bootstrap still exports it as a function. This causes 'tup
version' from the bootstrapped tup to show garbage. We can re-use
link.sh from the actual tup build to make sure we produce tup_version in
the same manner.

Fixes #384