Use a TAILQ for file_info lists.
This just makes it so that file accesses are processed in the same order
that they are used in the subprocess, though that shouldn't actually
affect the end result since things end up in trees. Mostly I found this
confusing while debugging something else that the order was inconsistent
with what I was expecting.