Add memory pools.
In one performance test, where many commands are run in parallel and
each with thousands of inputs, the millions of tiny malloc & frees for
things like pel_entry and tent_tree added up to a significant portion of
the total runtime (roughly 10%). Using memory pools for these numerous
small & fixed allocations removes much of this bottleneck.

The pools are thread local so that they can be lock-free. Only the
global list of allocations is locked, which is only needed for running
in valgrind.