Added t6085 - Make sure a resurrected file appears in a wildcard.
Since the file to be resurrected was in the gen_delete_root, it would
get skipped for wildcards. We had to check that files in the delete root
weren't in the modify list in other places like in nl_add_path and in
delete_files(), so it makes sense instead to simply move the deleted
files out of gen_delete_root and into a separate tree, now save_root.