Use fuse3 by default on Linux
Aside from a few minor API changes, the only main issue is that we don't
always get a mknod() to add the file mapping. The open() function now
also needs to create a mapping if we're writing a file, making sure we
skip hidden & excluded files.

Using fuse3 is now the default on Linux, but fuse2 still needs to be
supported in OSX. Linux can also use fuse2 by bootstrapping with
TUP_SERVER=fuse in the environment and setting CONFIG_TUP_SERVER=fuse in
tup.config.