Don't require a .tup directory for 'tup generate'
When 'tup generate' was originally created, a .tup directory was not
required or expected. However, this broke when Windows support was added
in 07d2e69fe. We can work around that instead by setting the tup_wd in
open_tup_top() if it isn't already set in find_tup_dir().