Fix Windows tmpfile() wrapper to use getenv(TMP) instead of .tup/tmp
So tmpfile() is definitely broken on Windows and we have to wrap it,
because it tries to open files in the "root directory" (according to the
docs), which can result in Permission Denied when using a VS developer
shell.

However, we don't want to need a .tup directory for creating tmp files
for 'tup generate', so instead of creating them in .tup/tmp, we'll put
them in getenv(TMP).