win32: Even more better directory test.
This somewhat reverts 4e28c2bfdbd54a653b9482bf4a5e85862538b0a7, though I
think that was intended to fix the same issue. t5092 was intermittently
failing on Windows, because depending on the scheduling of the two
run.sh scripts, sometimes it will get a write event on 'ignoredir', even
though it's a directory and we shouldn't be getting write events for it.
Apparently the test created in the other commit doesn't always work, so
we should also check for CreateOptions' FILE_DIRECTORY_FILE as well.

The other flaky test is t5091, which revealed that the
GetFileInformationByHandle wasn't actually returning successfully,
partly because we were using the FileHandle incorrectly and partly
because it fails with invalid permissions (not sure why on the latter
part, seeing as how it was just opened).