Database version 18 - Add a ^t flag for transient outputs.
The ^t flag marks the output files of a command as "transient".
Transient files are deleted after all downstream commands have been
executed. The sole benefit of doing this is to save disk space, which
can help for specific workloads.

The database change is required in order to add the transient_list
table. This keeps track of commands that must be re-run when command
strings change. Simply putting it in the modify list is insufficient,
since the command will almost always be in the modify list due to
missing outputs. Additionally, the transient list keeps track of output
files that haven't yet been removed due to a partial update or build
failure/canceled build.

Fixes #405.