Add checks when adding/removing nodes to the graph lists
When working on the graph code it's too easy to try to add something to
a list when it's already on one, or remove it from a list that it's not
on. If one of those bugs creeps in, it can manifest itself with weird
issues (like a command being run twice) that are hard to identify.