Added t2219 - including a lua file shouldn't break Tupfile variables.
Including any lua file caused the parser's vardb to point to the lua
parser's external vardb, and any variable lookup went through there
first before checking tup's own variables. Part of the problem is that
tup's vardb gets converted into lua variables upon parsing a lua file,
which means that afterward the lua parser essentially owned all
variables. So it was not possible to, for example, create CFLAGS in a
Tupfile, then include a lua file, and then modify the CFLAGS in the
Tupfile.

This change makes it so that the 2-level vardb only functions while a
lua file is being parsed. After parsing is complete, all lua variables
(except the defaults) are converted back into the parser vardb.