Add environment variable logging to --debug-logging
A --debug-logging update wasn't showing anything when an environment
varaible changed in the debug.log, though the node would show up in the
.dot graphs at least. Additionally, starting the server during the
parser stage, but not actually parsing anything, would cause the cwd to
change to '/', and then tup would fail to write out the update.dot graph
since it was using a relative path. These now use full paths instead.