Added t3090, t3091 - test for 'tup varsed' with 'tup generate'.
The generate command needs to create the vardict file alongside the
script in order for 'tup varsed' to work in the build script. Also, the
build script needs to export the environment variable to point to the
vardict file.

Normally the vardict file is in the .tup directory, but since that
doesn't exist during 'tup generate', it just gets created at the same
level as the build script itself.