win: Prevent tup from running if ConEmu is using DLL injection.
The ConEmu Windows Terminal has an option to do its own DLL injection.
If enabled, this interferes with tup's DLL injection for dependency
tracking, causing a "failed to inject dll" error.

This fix simply checks if the DLL is present in the tup process, and
provides a more helpful error message.