Parser version 14: Change variants to use explicit paths rather than an overlay.
Instead of overlaying the objdir with the srcdir while running commands
in a variant, we can use the fact that we usually know when a file is
generated or not to output the correct path when using %-flags. In other
cases, we can use TUP_VARIANTDIR to correspond to the directory of the
variant (or '.' for the non-variant case).