Support !tup_preserve on windows
The !tup_preserve logic was based around symlinks, and did the relative
path generation in the parser. On Windows we need to do file copies, and
the symlink-oriented relative paths aren't helpful there. Instead we can
move the relative-path variant logic into server_symlink, and Windows
then has enough correct information to determine the file copy paths.

The server_symlink() for fuse & ldpreload was moved into symlink.c,
since it is shared between the two and is large enough now that
copy/pasting it is undesirable.