Added t6086 - make sure 'deleting files' progress bar can reach 0.
After splitting out the save_root, the progress bar for removing files
could fail to reach 0 if files were removed and other files were
resurrected. This is because the count fields weren't accurately handled
in split_roots().

Now tent_tree always has a count field, so it doesn't rely on calling
the correct functions every time to make sure the count is up-to-date.