Re-work how "ghost" directories are parsed.
If we blow away a directory that was writing files to another directory,
commit 0c985485b originally handled this by calling parse() on the ghost
directory to get the generated files in other existing directories to be
deleted. Instead we can add those generated files to gen_delete_root by
querying based on the srcid for ghost directories, which helps fix a
problem counting directories.