Adjust progress bar for small numbers.
If the progress bar has 3 elements, then at 1 unit complete (33%) the
progress gets rounded down and fills out 0. Similarly at 66%, it fills
out 1, and then jumps to 3 after the 3rd element completes. For cases
where the progress bar is equal to the number of elements, we can
similarly set the fill value to be the number of completed elements.