Correct the progress bar count.
In some circumstances the progress bar count was incorrect, and counted
more nodes than were actually parsed. When parsing was complete, it
would report less than 100% complete.

Now there's a check to make sure that the previous progress bar was
completed, which prints an error message normally, but causes tup to
exit if running with TUP_VALGRIND=1 for testing purposes.