Added t2222 - make sure gitignore files aren't converted unnecessarily.
If a .gitignore file ends up in the modify list, it wasn't being removed
from save_root when the .gitignore directive was processed. This would
cause it to become a normal file after all gitignore files were
generated. Then on a subsequent parsing, tup would see that the
gitignore file was a normal file with a .gitignore directive, and
convert it back to a generated file. Each further parse would ping-pong
back & forth between the two.

When the .gitignore directive is reached, we need to remove it from both
the gen_delete_root as well as the save_root, similar to when outputs
are processed.