Added t8105 - test for progress bar when variant dir is deleted.
If a variant dir is deleted with a ghost node inside of it, we end up
with a ghost directory in the create list. This ends up in the parser
graph, which then gets converted back to a normal directory as a result
of the variant processing in get_rel_tent(). Now the node is expected to
be counted, but it didn't get added to num_nodes.

We already post-process the nodes in the graph to remove inactive
variant nodes, so we can check for the uncounted nodes there as well.