Update Lua to 5.4.3
The tup-lua.patch mostly applied as-is, with the main exception of the
argument adjusting in append_assignment. VINDEXED became split into
VINDEXED, VINDEXI, and VINDEXSTR - the latter two have 1 reg to be
added, while VINDEXED still has 2. Additionally, the condition that is
checked in freereg() was changed, which is now mimiced in unfreereg().

Unfortunately I haven't been able to get the automatic tostring() to be
called when you do print(table). It looks like we would need to set a
default metatable on all tables somehow, which is likely to cause other
issues. However, you already need to call tostring in certain instances
(such as string concatenation), so it seems reasonable to require it in
this one.