Parser version 15: Add sticky links back for normal non-order-only files.
A while back (8b60394ee) sticky links on normal files were no longer
added because sticky links were going to be propagated up the tree in
order to allow transitive dependencies. However, transitive dependencies
are now handled by PDAG generation, so that concern no longer applies.

It can be useful to know which inputs to a command were listed in the
Tupfile, such as for determining which input to a command is the "main"
file. The need for this comes from compiler and editor plugins that need
to know the command-line used to compile files. If we have a sticky link
from foo.c to the compilation command, these can easily be retrieved
from the tup database.