Add 'tup compiledb' to create compile_commands.json for files marked with ^j
This tries to generate the compilation db by first making sure all
Tupfiles have been parsed. Then it goes through every command that was
marked with the '^j' flag and produces a compile db entry for every
sticky link. Now that there are sticky links on normal files again, this
ensures that we can create an entry for foo.c, but not any of the
included headers.