Added t4218 - make sure tup doesn't fail if stdin is closed.
Tup closes stdin and dups a /dev/null fd over it before shelling out to
subprocesses. However, if stdin was already closed by the process
calling tup, then one of master_fork's socketpair descriptors would
become 0. Duping /dev/null over part of the socketpair would close the
pipe and cause tup to quit.

The workaround is to open a temporary file descriptor, so the socketpair
is never one of the standard descriptors.