Add help when '-h' or '--help' is passed in.
Most other programs give help when -h or --help is given as an argument.
For programs with lots of subcommands like tup, 'tup subcommand -h'
should give help on that specific subcommand.

The initial set of subcommands is copied as a shortened version of the
man page. Ideally they would both pull from the same primary source of
information, but this is a reasonable start. Not all subcommands appear
in the man page or help yet either.

Fixes #420.