Added t4219 - test for 'tup refactor' with an exclusion.
Apparently calling tup_db_create_node(..., TUP_NODE_GHOST) when the node
already existed as a ghost would cause a bunch of unnecessary SQL
queries that were re-setting already-set parameters and adding it to the
modify list. This caused a 'tup refactor' with an ^exclusion (which is a
ghost) to fail.