Added t5116 - add a new command dependent on a transient command.
When a transient command is run and its output no longer exists, if we
then add a new command that depends on the transient file, we only have
a sticky link pointing from the transient file to the new command.
However, the attach_transient_nodes() logic was only operating on normal
links, which means transient files only got pulled in if the command
executed at least once first.

The fix is to pull in transient nodes if a sticky link exists, so that
when we run any command that depends on a transient file, we can ensure
that the transient file exists when the command runs.