Database version 19 - add mtime_ns for nanosecond timestamps.
Prior to this version, tup only tracked mtime in 1-second resolution.
This required a bunch of hacks in the test cases, otherwise a test like
the following would fail to recognize that a Tupfile had changed:

echo 'some rule' > Tupfile
update
echo 'new rule' > Tupfile
update

Since this would almost always complete within the same 1s interval, the second update would see the Tupfile had the same timestamp as the last update, and not re-parse it. In test cases this was worked around by throwing 'tup touch' commands virtually everywhere, but that is just a confusing hack.

If we also track nanosecond times as well, the hacks aren't necessary
and the test cases can be much more straightforward.