Remove extraneous 'touch' commands in test scripts.
These were originally 'tup touch's used to tell tup that a file had
changed when quickly running updates back-to-back inside of the 1s
resolution that the tup db tracked. Now that we use ns resolution, these
extra 'tup touch' commands are largely unnecessary.

There are still a few cases where we want to 'tup touch' to update the
database directly and skip the scanning phase in order to test certain
behavior. However, we no longer need to 'tup touch' in the vast majority
of the cases where we can now expect scan to detect the differences
properly.