Update t4047 to ignore fds opened outside of the test case.
If you open an fd before running the test (eg: the test-running
environment has a log opened), then t4047 would fail because it would
see an open fd.

We can instead compare the current open fds with the ones opened while
running under tup to see if tup is opening any extraneous fds. All we're
trying to check here is that tup isn't leaking fds to the child
processes unexpectedly.