Added t6087 - broken update with tup internal error: tup_entry_rm called on tupid X, which still has refcount=1
We don't normally add dependencies on directories to commands, but if a
command has a dependency on a ghost that later becomes a directory, we
can end up with the command in the create graph. This is also normally
fine, since we only process directories in the create graph.

However, if the ghost becomes a directory at the same time the command
is deleted, then the command ends up in the cached stickies of the
directory node, and when we try to remove the command & output file, we
get an internal refcount error.

This was found when trying to compile a really old version of tup and
then going back to the master branch.