win32: Remove the WaitForInputIdle() calls.
These calls always seem to fail (returning -1), or in the case of
GhostScript, time out. The latter is problematic since the timeout
given to the function was infinite, so tup would appear to hang.

I'm not sure what causes GhostScript to fail with the timeout rather
than failing immediately like everything else, but given the error
message (ERROR_NOT_GUI_PROCESS), it seems unlikely to be something we
actually want to use.

I think this went unnoticed for so long because WaitForInputIdle is a
rare Windows function that returns 0 on success, while most others
return 0 for failure. So the test to see if it failed was actually
incorrect, and since it always failed, things kept working. Until, that
is, the GhostScript timeout hang.