In 'tup graph --combine', sort combined nodes alphabetically.
For some reason, the Github automation finds foo.h in the filesystem
before bar.h on the initial scan. Tup was combining nodes based on
tupid, so the test case failed since the check was assuming bar.h would
be found first. We can instead sort the nodes alphabetically, so the
graph should be the same regardless of how nodes were entered into the
db.