Change internal representation of node variables to be a "%123t" string.
We can treat a node varaible as a string that is basically the tupid,
with a string like %123t, where 123 is the tupid. These strings can be
passed around in variables and such internally, and only evaluated at
the endpoint where they are needed.

In future commits we can make TUP_VARIANTDIR a node variable, so that it
can be evaluated differently depending on whether or not its in a string
or an input/output.