Added t5117 - make sure mtimes are updated in the db after restoring outputs.
If we restore an output because a ^o rule failed, the rename() affects
ctime, so we need to update the db with the new time. Otherwise if the
rule gets removed before successfully building, the output file will
have a different mtime from what we think and get resurrected when it
shouldn't have been.