Added t7061 - make sure ^o rules failing with autoupdate don't cause infinite loops.
When a ^o rule fails, the outputs get restored, and the monitor picks
up the rename() call. This shows up as an IN_MODIFY event due to
ephemeral_event() detection, and triggers another autoupdate.

Instead we can ignore IN_MODIFY events on generated files when deciding
on autoupdates. If a user manually removes an output, presumably they
don't want it immediately auto-updated and re-created anyway.