Added t4221 - test for stderr with failing ^o command.
This broke with commit af5ae9ab27 - if a command with ^o failed and
hadn't produced an expected output yet, the lstat() would fail and tup
would only show that error instead of the command's stderr. If the
lstat() fails due to ENOENT we should just skip setting the mtime.