Added t5118, t8112 - correctly handle symlinks with outputs in other directories.
The link path was being calculated assuming the output file was created
in the directory being parsed. So if you used !tup_preserve with a ../%B
or similar output, the link path would be invalid (or point to the wrong
file). We already know the actual output though, so we can calculate the
link path relative to the output file itself.