Added t6088, t4223 - exclusion tests for generated files.
While testing that an exclusion could be used to ignore .gitignore
files, I ran into a problem with refcount being off when removing an
entry and causing an internal error. The issue was that I had
^.gitignore as the exclusion, which as a regular expression matched the
path '/.../test-gitignore/out' (the output file for the command). Tup
failed to create a new ignore file on top of the generated node since it
was still listed as an output for the command, resulting in the command
failing and the refcount not updating correctly. t6088 tests for this,
with the correct behavior being a more explicit error message.

t4223 tests for the original use-case, which was to ignore dependencies
on .gitignore files. Specifying ^/.gitignore as the ignore pattern
matches the gitignore files (not anything in a directory with
'gitignore' in it), and we can verify that .gitignore files are indeed
ignored with this pattern.