Added t6089 - test for an internal & explicit file write ignore.
Writing to a file inside a .git directory is ok (it just generates a
warning about writing to a hidden file). However, if you also explicitly
ignore it with an exclusion like ^\.git, then tup checks that first and
fails to find a pel pointer and returns an internal error.

Instead we can do the internal checks first, and only if those pass we
can check the exclusions.

Fixes #468.