Added t2234, t2235, t2236 - Only add lua vars to vardb once all lua parsing is complete.
t2234 checks if a lua file includes another lua file, the internal
variables (like _G, the tup table, etc) shouldn't get added to the vardb.

t2235 checks to make sure we can add to CFLAGS that are specified in a
Tuprules.lua from a Tupfile.lua.

t2236 checks to make sure that global variables in Tupfile.lua's don't
cross-pollute. This could be an issue if a single lua_State is used in
the future.