Put tupfile structs in a stack to grab for lua parsing.
Instead of setting the tup.XYZ functions to refer to the current Tupfile
being parsed everytime we load a Tupfile.lua file, we can just grab the
currently parsed tupfile structure from the parser. This goes in a stack
since sometimes we start parsing a Tupfile and then switch to parsing
another directory if we use that directory's files as inputs.