Parser version 16 - use a single Lua state for all Tupfiles.
When parsing a Lua-based Tupfile, a new lua_State was created for each
file, which had to parse all the tup definitions and builtin rules every
time. Additionally, when switching between regular Tupfile parsing and
lua parsing, the vardb had to be copied back and forth so you could do
things like append to CFLAGS both with and without lua, but have it
affect the same variable.

Using a single lua_State, we can instead parse the builtin rules and
definitions only once during tup initialization. A simple sandboxing
approach is used to isolate "global" variables into a per-Tupfile
tupvars structure so Tupfiles don't cross-pollute each other. Regular
Tupfiles now use the lua state to create variables as well instead of
our own vardb struct. So if a regular Tupfile sets CFLAGS, it is already
in lua and doesn't need to be copied in and out when a lua file is
parsed.