Lua parser - set __tostring and __concat for variable tables.
When creating variables in Tupfiles or using the += operator in
Tupfile.lua, the corresponding lua table now has a metatable assigned
that will expand the table when using string concatenation (eg: 'foo' ..
CFLAGS) or when __tostring is called (for things like print() or error()
or whatever).

Note that if you create a table manually, like:

CFLAGS = {'-DFOO', '-DBAR'}

Then the metatable is not automatically assigned to it. You can still
use the already overridden 'tostring' to have tostring(CFLAGS) expand
into '-DFOO -DBAR', but concatenation won't work by default. To work
around this you could assign the tup_table_meta manually, or just always
assign with +=