Added t5119 - Return EPERM when trying to write to a directory that we don't have permission to.
With the FUSE overlay, file writes are done in a directory that tup can
always write to. This could cause the program to run differently in tup
than in a shell, because in tup the write will always appear to succeed.
Instead we can test that it is actually possible to write to the
directory when creating files.