Added t5120 - test for ignoring outputs outside tup with full_deps enabled.
Ignoring an output causes tup to create a placeholder node for the file
so that a scan won't see the new file and think it needs to re-parse a
directory. However, it was always creating this file as a TUP_NODE_FILE,
which wasn't correct when it was outside the tup hierarchy (all files
outside of the hierarchy are expected to be ghosts).

This was causing an 'Unable to find parent node' error on subsequent
updates, which was difficult to recover from. The fix is to make the
placeholder nodes TUP_NODE_GHOST if they are outside of the tup
hierarchy.