fuse: Treat renames to '.fuse_hidden' files as unlinks.
In my Arch VM, sometimes a file we be renamed to '.fuse_hiddenXYZ' when
the subprocess removes the file, and tup won't get a subsequent unlink()
call on the .fuse_hidden file. This leads tup to think that the file was
created by the subprocess, when obviously it isn't the case.

We can work around this by treating a rename() call to a .fuse_hidden
file as if it was an immediate removal.