Work around low-resolution filesystem timestamps in Linux.
Even though tup uses nanosecond timestamps as of b5e39d33, the actual
resolution of timestamps in ext4 is based on the kernel tick rate, which
means something like 3-4ms on a default Arch or Ubuntu install.
Unfortunately that large of a time delta is enough for tup to run and
new files to be updated within a test case to setup for the next tup
call, meaning the test run won't see that the files have been updated.

Increasing CONFIG_HZ seems to help, but anyone else that wants to run
the test cases on their Linux machine would need to do the same. Another
option is to add 'tup touch' commands back into test cases, but if those
are missing it results in intermittent test failures, which are
annoying. The option implemented here is to pass in the value of
CONFIG_HZ into tup when running a test case on Linux, so tup can sleep a
full tick after executing so that future filesystem changes will have
updated timestamps.

Ideally Linux would have higher resolution filesystem timestamps, but
that isn't currently the case.