Added t8115, t8116 - tests for things in variants becoming directories.
If we have a generated file that is built in a variant directory, and
later the user creates a directory of the same name of the generated
file, we have to remove the generated file in the variant directory
before trying to parse it. Similarly, t8116 tests the case where there
is a generated directory in the variant and then becomes a real
directory.