Allow overriding CC on macos & freebsd.
Other platforms allow overriding CC in the bootstrap build by setting CC
in the environment, but macos & freebsd were setting it back to clang.
Similarly, if CONFIG_CC is set in the tup.config, the actual tup build
for macos & freebsd would ignore that value and just use clang.

Now the CC environment variable for bootstrapping and CONFIG_CC for the
tup build should work for setting the compiler in all cases.

Fixes #497.