Added t8117, t8118, t8119 - support for variants in subdirectories.
Variants were originally limited to be subdirectories of the top-level
directory (ie: siblings of the .tup directory itself). This was largely
to simplify the already complex problem of creating variants in the
first place, but now it appears to be an artificial limitation.

Fixing this mostly involved removing the safeguards that prevented
creating variants in subdirectories, as well as better support for
printing multi-part variant directories in print_tup_entry().

You still can't create a variant inside another variant, which is what
t8118 and t8119 check for. But you can create a single "builds"
directory and create all variants inside of that, for example.