Add initial support for the 'tup variant' command on Windows.
On other OSes, the 'tup variant' command creates a symlink to the config
files, but since that doesn't work easily on Windows we just copy the
file instead. We have to calculate a slightly different path to make
that work, and adjust the test cases to account for the fact that the
variant won't be updated when the source config file is modified.