Direnv support
Direnv <https://direnv.net> is a tool for tracking your local developer
environment. With the rise of Nix popularity & its hooks for local Nix
development, this tool has become quite popular, but it caches changes
in a local directory that makes Tup noisy.

This commit is a duplicate of another proposed patch, however, unlike
that patchset, this is standalone.