Attempt to continue working without namespacing if apparmor denies user namespaces.
If apparmor is configured to deny unprivileged user namespaces, the
unshare(CLONE_NEWUSER) call will still work, it just won't have any
privileges (like CAP_SYS_ADMIN) that we need. This will manifest itself
in the deny_setgroups() function when we try to open /proc/pid/setgroups
for write.

When the unshare() call itself fails, we try to continue on without
namespacing enabled. This allows at least for diminished functionality
where .tup/mnt paths are visibile to subprocesses. However, when the
deny_setgroups() function fails, tup was just quitting outright. It
would be better if we could continue in the diminished capacity as if
the unshare() call itself failed.

Unfortunately, since unshare() isn't failing, we're already in a new
namespace by the time we know that it doesn't have the capabilities we
need. Therefore we need to close out the child master_fork process and
create a new one in the original namespace. This is the reason for the
funkiness of writing out a special restart token ("2").

To prevent these error messages from showing up in the first place, one
could set TUP_NO_NAMESPACING=1 in the environment to force tup to run
without trying namespacing in the first place. Another option is to
update the apparmor configuration for tup. For example, Ubuntu comes
with a default tup configuration in /etc/apparmor.d/tup that contains:

---
abi <abi/4.0>,
include <tunables/global>

profile tup /usr/bin/tup flags=(unconfined) {
  userns,

  # Site-specific additions and overrides. See local/README for details.
  include if exists <local/tup>
}
---

The path (here, /usr/bin/tup) needs to match the path of the actual tup
binary. So if building tup from source, these paths won't match. You
could add another profile in the same file or create a new apparmor
profile like so:

---
profile localtup /path/to/local/tup/tup flags=(unconfined) {
  userns,

  # Site-specific additions and overrides. See local/README for details.
  include if exists <local/tup>
}
---

Fixes #502.